Apply review comments
- Wrap error when calling HasPendingRebootDBParams
- Use a single channel containing a struct with both the bool and error
  in HasPendingRebootDBParams
- Move all new DB-related methods to terraform/db_operations.go
